requirements.txt
numpy~=1.22.2
pandas~=1.4.2
opencv-python~=4.5.4.60
albumentations~=1.2.1
tqdm~=4.62.3
wandb~=0.12.21
PyYAML~=6.0
matplotlib~=3.4.3
albumentations_experimental==0.0.1
